Add divider unit and integrate into pipeline
- Introduced a new divider unit (`div_unit`) to handle division and remainder operations for RV32M extension, supporting DIV, DIVU, REM, and REMU instructions.
- Updated `exec_pipe1` to include outputs for the divider unit and modified routing logic to handle division operations.
- Enhanced the instruction issue arbiter (`iq_pipe1_arbiter`) to prioritize division requests alongside memory and multiplication requests.
- Modified load/store unit (`lsu_shell`) to include verbose logging for division-related operations.
- Updated instruction stage (`stage_is`) to classify division instructions correctly.
- Expanded test suite in `run_all_tests.py` to include basic and helper path tests for division.
- Added new assembly test files (`test_div_basic.s` and `test_div_helper_path.s`) to validate division functionality.
- Adjusted the build process to accommodate new tests and ensure proper architecture flags are set for division-related tests.

+++ benchmarks/dhrystone/dhrystone_main.c
@@ -224,9 +224,9 @@ int main (int argc, char **argv)
     Bench_Ipc_X1000 = 0UL;
   }
 
-  printf("BENCH CYCLES: %llu\n", Bench_Total_Cycles);
-  printf("BENCH INSTRET: %llu\n", Bench_Total_Instret);
-  printf("BENCH IPC_X1000: %lu\n", Bench_Ipc_X1000);
+  printf("BENCH CYCLES: %u\n", (unsigned)Bench_Total_Cycles);
+  printf("BENCH INSTRET: %u\n", (unsigned)Bench_Total_Instret);
+  printf("BENCH IPC_X1000: %u\n", (unsigned)Bench_Ipc_X1000);
 #ifndef AX7203_DHRYSTONE_EFFECTIVE_FIXED_RUNS
   printf("Microseconds for one run through Dhrystone: %ld\n", Microseconds);
   printf("Dhrystones per Second:                      %ld\n", Dhrystones_Per_Second);
@@ -237,9 +237,9 @@ int main (int argc, char **argv)
 #else
   for (;;) {
     board_delay_ms(250);
-    printf("BENCH CYCLES: %llu\n", Bench_Total_Cycles);
-    printf("BENCH INSTRET: %llu\n", Bench_Total_Instret);
-    printf("BENCH IPC_X1000: %lu\n", Bench_Ipc_X1000);
+    printf("BENCH CYCLES: %u\n", (unsigned)Bench_Total_Cycles);
+    printf("BENCH INSTRET: %u\n", (unsigned)Bench_Total_Instret);
+    printf("BENCH IPC_X1000: %u\n", (unsigned)Bench_Ipc_X1000);
     printf("Microseconds for one run through Dhrystone: %ld\n", Microseconds);
     printf("Dhrystones per Second:                      %ld\n", Dhrystones_Per_Second);
     printf("DHRYSTONE DONE\n");
